pyproject.toml
[project]
name = "vibe-manager"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "httpx>=0.28.1",
]

[project.scripts]
vibe = "vibe.main:app"

[tool.hatch.build.targets.wheel]
packages = ["vibe"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
